Fix scope closing bug
Fixes bug in server tracing handler where server span was being closed
early.

Instead of using the try-with-resources block - which seems to close the
scope too early (possibly because the code inside it is async?),
creating a SpanInScope and closing it on the callback that's invoked
before the server response is sent.

Add Ok-Http MockServer dependency for testing.