Make guice components singletons
Currently the guice components are not singletons.  Besides the unecessary instantiation, this also breaks things like the CurrentSpanCustomizer pattern where each CurrentSpanCustomizer was receiving a new tracer.  Since each tracer has its own recorder, additional tags are always added to a new span instead of the current span.

In addition to adding the singleton scopes this also exposes `Tracing`, `Tracer` and `SpanCustomizer` on the guice registry for convenience.