Refactors tracing handler to use HttpServerParser
This commit introduces an implementaton of HttpServerParser and
refactors the server tracing handler such that span name customization
is done via the HttpServerParser instead of directly in the tracing
handler.

In order to make this work, needed to add a wrapper around the standard
Ratpack Response so that things like the PathBinding and original
ServerRequest are available to the HttpServerParser.

Closes #51